Add security config
