+++ src/tests/routers/movie.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter
 from fastapi import Path, Query, Depends
 from fastapi.responses import JSONResponse
 from typing import List
-from config.database import Session
+from config.db import Session
 from models.movie import Movie as MovieModel
 from fastapi.encoders import jsonable_encoder
 from middlewares.jwt_bearer import JWTBearer
